@@ -12,64 +12,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/pwa": "^0.1100.2",
-    "@angular-devkit/schematics": "~11.0.2",
-    "@angular/platform-browser": "~11.0.2",
-    "@angular/platform-browser-dynamic": "~11.0.2",
-    "@angular/animations": "~11.0.2",
-    "@angular/cdk": "^11.0.2",
-    "@angular/common": "~11.0.2",
-    "@angular/compiler": "~11.0.2",
-    "@angular/core": "~11.0.2",
-    "@angular/forms": "~11.0.2",
-    "@angular/material": "^11.0.3",
-    "@angular/router": "~11.0.2",
-    "@angular/platform-server": "~11.0.2",
-    "@angular/service-worker": "~11.0.2",
-    "@ngx-translate/core": "^11.0.1",
-    "@nguniversal/module-map-ngfactory-loader": "9.0.0-next.9",
-    "zone.js": "~0.11.3",
-    "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.3.1",
-    "bootstrap-material-design": "^4.1.2",
+    "@angular-devkit/schematics": "~13.1.2",
+    "@angular/animations": "~13.1.1",
+    "@angular/cdk": "^13.1.1",
+    "@angular/common": "~13.1.1",
+    "@angular/compiler": "~13.1.1",
+    "@angular/core": "~13.1.1",
+    "@angular/forms": "~13.1.1",
+    "@angular/material": "^13.1.1",
+    "@angular/platform-browser": "~13.1.1",
+    "@angular/platform-browser-dynamic": "~13.1.1",
+    "@angular/platform-server": "~13.1.1",
+    "@angular/pwa": "^13.1.2",
+    "@angular/router": "~13.1.1",
+    "@angular/service-worker": "~13.1.1",
+    "@ngx-translate/core": "^14.0.0",
+    "bootstrap": "^5.1.3",
+    "bootstrap-material-design": "^4.1.3",
     "bootstrap-notify": "^3.1.3",
     "core-js": "^3.2.0",
     "html2canvas": "^1.0.0-rc.3",
     "jquery": "^3.5.1",
-    "jspdf": "^1.5.2",
+    "jspdf": "^2.5.0",
     "ngx-bootstrap": "^5.1.1",
     "oidc-client": "^1.7.1",
     "perfect-scrollbar": "1.1.0",
-    "popper.js": "^1.15.0",
-    "rxjs": "~6.6.3",
-    "rxjs-compat": "~6.6.3",
-    "web-animations-js": "^2.3.2",
-    "xlsx": "^0.15.0",
+    "popper.js": "^1.16.1",
+    "rxjs": "~7.5.1",
+    "rxjs-compat": "~6.6.7",
     "save-dev": "0.0.1-security",
-    "tslib": "^2.0.3"
+    "tslib": "^2.3.1",
+    "web-animations-js": "^2.3.2",
+    "xlsx": "^0.17.4",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.2",
-    "@angular/cli": "~11.0.2",
-    "@angular/compiler-cli": "~11.0.7",
-    "@angular/language-service": "~11.0.7",
-    "@types/jasmine": "3.5.0",
-    "@types/jasminewd2": "2.0.8",
-    "@types/node": "13.1.1",
-    "codelyzer": "5.2.1",
-    "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "^5.1.0",
+    "@angular-devkit/build-angular": "^13.1.2",
+    "@angular/cli": "~13.1.2",
+    "@angular/compiler-cli": "~13.1.1",
+    "@angular/language-service": "~13.1.1",
+    "@types/jasmine": "3.10.2",
+    "@types/jasminewd2": "2.0.10",
+    "@types/node": "17.0.4",
+    "codelyzer": "6.0.2",
+    "jasmine-core": "3.10.1",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "^6.3.9",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.5.1",
-    "typescript": "4.0.2"
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.7.0",
+    "typescript": "^4.4.2",
+    "webpack": "^5.65.0"
   },
   "optionalDependencies": {
-    "node-sass": "^4.9.3",
-    "protractor": "~5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1"
+    "eslint": "8.5.0",
+    "node-sass": "^7.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.4.0"
   }
 }
