@@ -1,5 +1,6 @@
 import { Component, OnInit, ViewChild, Input, EventEmitter, Output, ElementRef } from '@angular/core';
 import { Router } from '@angular/router';
+import { MatFormField } from '@angular/material/form-field';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatSort } from '@angular/material/sort';
 import { MatPaginator } from '@angular/material/paginator';
@@ -8,11 +9,14 @@ import { MemberRegistrationGridModel } from './Models/memberRegistrationGrid.mod
 import { PaginationService } from '../Shared/PaginationService';
 
 @Component({
-    templateUrl: 'allMemberRegistration.component.html'
+  selector: 'allMemberRegistration',
+  templateUrl: './allMemberRegistration.component.html'
 })
-export class AllMemberRegistration implements OnInit
-{
+export class AllMemberRegistration implements OnInit {
+
   data: MemberRegistrationGridModel[];
+  dataSource: MatTableDataSource<MemberRegistrationGridModel>;//= new MatTableDataSource<MemberRegistrationGridModel>();
+
   errorMessage: any;
 
   @ViewChild(MatSort, { read: ElementRef, static: false })
@@ -21,47 +25,44 @@ export class AllMemberRegistration implements OnInit
   @ViewChild(MatPaginator, { read: ElementRef, static: false })
   paginator: MatPaginator;
 
-    displayedColumns: string[] = ['MemberId', 'MemberNo', 'MemberName', 'Contactno','PlanName', 'SchemeName', 'JoiningDate', 'EditAction', 'DeleteAction'];
-    dataSource = new MatTableDataSource<MemberRegistrationGridModel>();;
-    @Input() totalCount: number;
-    @Output() onPageSwitch = new EventEmitter();
+  displayedColumns: string[] = ['MemberId', 'MemberNo', 'MemberName', 'Contactno', 'PlanName', 'SchemeName', 'JoiningDate', 'EditAction', 'DeleteAction'];
 
-    constructor(private _Route: Router,private memberregistration: MemberRegistrationService, public paginationService: PaginationService) { }
+  @Input() totalCount: number;
+  @Output() onPageSwitch = new EventEmitter();
 
-    ngOnInit() {
-        this.memberregistration.GetAllMember().subscribe(
-            allMember => {
-                this.data = allMember
-                this.dataSource = new MatTableDataSource(this.data);
-                this.dataSource.sort = this.sort;
-                this.dataSource.paginator = this.paginator;
-            },
-            error => this.errorMessage = <any>error
-        );
-    }
+  constructor(private _Route: Router, private memberregistration: MemberRegistrationService, public paginationService: PaginationService) { }
 
-    Delete(MemberId): void 
-    {
-        console.log(MemberId);
-        if (confirm("Are you sure to delete Member ?")) 
-        {
-          this.memberregistration.DeleteMember(MemberId).subscribe
-            (
-            response => {
-              if (response.StatusCode == "200") 
-              {
-                alert('Deleted Member Successfully');
-                location.reload();
-              }
-              else {
-                alert('Something Went Wrong');
-                this._Route.navigate(['/Member/All']);
-              }
+  ngOnInit() {
+    this.memberregistration.GetAllMember().subscribe(
+      allMember => {
+        this.data = allMember;
+        this.dataSource = new MatTableDataSource(this.data);
+        this.dataSource.sort = this.sort;
+        this.dataSource.paginator = this.paginator;
+      },
+      error => this.errorMessage = <any>error
+    );
+  }
+
+  Delete(MemberId): void {
+    console.log(MemberId);
+    if (confirm("Are you sure to delete Member ?")) {
+      this.memberregistration.DeleteMember(MemberId).subscribe
+        (
+          response => {
+            if (response.StatusCode == "200") {
+              alert('Deleted Member Successfully');
+              location.reload();
             }
-            )
-        }
-      }
-      applyFilter(filterValue: string) {
-        this.dataSource.filter = filterValue.trim().toLowerCase();
-      }
+            else {
+              alert('Something Went Wrong');
+              this._Route.navigate(['/Member/All']);
+            }
+          }
+        )
+    }
+  }
+  applyFilter(filterValue: string) {
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+  }
 }
